requirements.txt
django==4.1
djangorestframework==3.13.1
pytz==2021.1
django-soft-delete==0.9.21
django-safedelete==1.3.1
django_filter==22.1